Fix type hints
